feature: move componente Header para dentro de app.vue
